backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.7.1
Flask-CORS==5.0.1
Flask-Mail==0.10.0
python-dotenv==1.0.1
marshmallow==3.23.1
bcrypt==4.2.1
